Fix getid handler (update for wikka's new directory layout)
This allows /getid to be appended to a wiki page URL, and the resulting
HTML contains the <wiki-response> stuff with the page id. This is used by
the fbdoc tool when downloading pages from the wiki.

Nowadays, the handler must be placed in a directory of the same name, like
all the others. It's the same as for actions/.